requirements.txt
Flask==2.3.3
yt-dlp==2023.9.24
gunicorn==21.2.0
python-dotenv==1.0.0